pyproject.toml
[project]
name = "travai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.21.0",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langgraph-api>=0.4.28",
    "langgraph-checkpoint-sqlite>=3.0.2",
    "langgraph-cli[inmem]>=0.4.8",
    "memvid-sdk>=2.0.148",
    "pymupdf>=1.26.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
